agent-coding-loop/pyproject.toml
[project]
name = "coding-agent"
version = "0.1.0"
description = "Coding agent with context budget management"
requires-python = ">=3.10"
dependencies = [
    "tiktoken>=0.5.0",
    "lancedb>=0.18.0",
    "numpy>=1.26.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/coding_agent"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
